Well on the way to completely messing up the STEP code, just by flipping the Y coord!

+++ src/hid/gtk/vertex3d.h
@@ -1,13 +1,13 @@
 typedef struct
 {
-  float x;
-  float y;
-  float z;
+  double x;
+  double y;
+  double z;
   int id;
 
   /* STEP crap - to hell with encapsulation */
   step_id vertex_identifier;
 } vertex3d;
 
-vertex3d *make_vertex3d (float x, float y, float z);
+vertex3d *make_vertex3d (double x, double y, double z);
 void destroy_vertex3d (vertex3d *v);
